requirements.txt
alembic==1.15.2
Authlib==1.3.2
blinker==1.9.0
cachelib==0.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
cryptography==45.0.6
Deprecated==1.2.18
dnspython==2.7.0
email-validator==2.2.0
Flask==3.1.0
Flask-Caching==2.1.0
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
google-api-core==2.2.25.1
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.1
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.5.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
Pygments==2.19.2
pyodbc==5.2.0
pyparsing==3.2.3
python-dotenv==1.0.0
requests==2.32.5
rich==13.9.4
rsa==4.9.1
SQLAlchemy==2.0.40
typing_extensions==4.13.2
uritemplate==4.2.0
urllib3==2.5.0
waitress==3.0.2
Werkzeug==3.1.3
wrapt==1.17.3
WTForms==3.2.1
